requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
faiss-cpu
tiktoken
huggingface_hub
python-dotenv
